+++ CHANGELOG.md
@@ -5,9 +5,12 @@ Features:
 - Breadcrumbs: disableExpand
 - Remix Vite FormControl error fixed
 - Toast: Added support for toast render boundaries
+- useHotkeys: disable flag
 
 Fixes:
 
 - View: Fixed dividers incorrectly working with React.Fragments
 - useHotkeys: auto preventDefault
 - DropdownMenu: fixed double click handling
+- Slider: Float values
+- Slider: Tabular numbers